made changes to accommodate themes for multiple brands
