+++ packages/theme/src/index.ts
@@ -1,4 +1,4 @@
-export * from './colors';
+export * from './themes/atb-theme/colors';
 export * from './sizes';
 export * from './theme';
 export * from './typography';
