+++ packages/theme/src/themes/index.ts
@@ -1,0 +1,1 @@
+export {default as AtBThemes} from './atb-theme/theme';
