+++ packages/theme/tools/create-css.ts
@@ -1,7 +1,8 @@
 import outputThemes from './create-theme';
 import outputTypography from './create-typo';
-
-Promise.all([outputThemes(), outputTypography()]).then(
+import { AtBThemes } from '../src/themes';
+Promise.all([outputThemes('atb-theme',AtBThemes), outputTypography()]).then(
   () => console.log('Written CSS files'),
   console.error,
 );
+
